bugfix
moving the .wav suffix into the temporary file name to avoid adding it
twice in the subsequent lines is incorrect because the stream_file
function requires the filename without the suffix.

It needs to be either stripped off for the playback, or not added in
the first place. I opted for the latter as this is closer to the
original behaviour anyway. Sorry for the problem.